move searchDropdown to a separate chunk
