+++ src/shared_components/TopBar/Search/Search.js
@@ -1,18 +1,11 @@
 // NPM
-import React, { Component } from 'react';
-import styled from 'styled-components';
-import { getSearchParams } from 'libs/location';
-import history from './../../../main/history';
-
-// COMPONENTS
-import { CrossIcon, SearchIcon } from '../../icons';
-import SemanticLocationControl from 'shared_components/Form/LocationAutoSuggest';
-import { geocodeByAddress, getLatLng } from 'libs/placesAutocomplete';
 import { disabled } from 'libs/colors';
-import { getAddress } from 'libs/search';
+import React, { Component, Suspense } from 'react';
+import styled from 'styled-components';
+// COMPONENTS
+import { SearchIcon } from '../../icons';
 
 // ACTIONS/CONFIG
-import { resetButton } from '../../../libs/styled';
 
 // STYLES
 const Wrapper = styled.div`
@@ -50,165 +43,24 @@ const IconButton = styled.span`
   }
 `;
 
-const Form = styled.form`
-  display: flex;
-  flex: 1;
-`;
-
-const Input = styled.input`
-  background: none;
-  border: none;
-  display: inline-block;
-  flex: 1;
-  font-family: inherit;
-  font-size: inherit;
-  font-style: inherit;
-  font-weight: inherit;
-  margin-right: 15px;
-  outline: none;
-  width: 100%;
-`;
-
-const ArrowWrap = styled.span`
-  color: #50a18a;
-  display: inline-block;
-  height: 16px;
-  margin-left: 7px;
-  margin-right: 3px;
-  width: 16px;
-`;
-
-const SubmitButton = styled.button`
-  ${resetButton({
-    fontWeight: '500',
-  })};
-  align-items: center;
-  color: #65afbb;
-  display: flex;
-  width: auto;
-  :active,
-  :focus {
-    outline: 0;
-    border: 0;
-  }
-`;
-
-const locationProps = {
-  inputStyles: {
-    height: '100%',
-    border: 0,
-  },
-  inputProps: {
-    placeholder: 'Where would you like to go?',
-    icon: 'false',
-    as: Input,
-  },
-};
-
 const suggestionStyle = {
   width: '80vw',
   maxWidth: '500px',
 };
 
+const SearchInput = React.lazy(() => import('./SearchInput'));
+
 // MODULE
 export default class DesktopSearch extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      search: '',
-      mode: 'text',
-      inFocus: false,
-      serviceType: props.searchParams.type || 'trip',
-      params: {},
-    };
-
-    this.handleInputChange = this.handleInputChange.bind(this);
-    this.handleSubmit = this.handleSubmit.bind(this);
-    this.onFocus = this.onFocus.bind(this);
-    this.onBlur = this.onBlur.bind(this);
-    this.handleLocationChange = this.handleLocationChange.bind(this);
-    this.handleSearchSubmit = this.handleSearchSubmit.bind(this);
-    this.searchWrapper = React.createRef();
-    this.inputRef = React.createRef();
-  }
+  state = { inFocus: false };
   onFocus() {
     this.setState({ inFocus: true });
   }
   onBlur() {
     this.setState({ inFocus: false });
   }
-  handleInputChange(ev) {
-    this.setState({ search: ev.target.value });
-  }
-  handleSubmit(ev) {
-    ev.preventDefault();
-    this.setState({
-      search: '',
-      mode: 'text',
-    });
-    this.inputRef.current.onReset();
-    history.push(`/results`);
-  }
-  handleLocationChange(address, serviceType, text) {
-    if (text) {
-      const params = {
-        city: undefined,
-        state: undefined,
-        countryCode: undefined,
-        lat: undefined,
-        lng: undefined,
-      };
-      this.setState({ text, params, address: null, serviceType }, this.handleSearchSubmit);
-      return;
-    }
-    geocodeByAddress(address).then(results => {
-      const result = results[0];
-      const searchParams = getSearchParams(address, result);
-      this.setState(
-        { address, params: searchParams, text: null, serviceType },
-        this.handleSearchSubmit,
-      );
-    });
-  }
-  handleSearchSubmit() {
-    if (this.props.toggleSearch && this.props.isMobile) {
-      this.props.toggleSearch();
-    }
-    const { text } = this.state;
-
-    const params = {
-      ...this.props.searchParams,
-      ...this.state.params,
-      topRightLat: undefined,
-      topRightLng: undefined,
-      bottomLeftLat: undefined,
-      bottomLeftLng: undefined,
-      text,
-      type: text ? 'trip' : this.state.serviceType || this.props.searchParams.type,
-      priceStart: undefined,
-      priceEnd: undefined,
-    };
-
-    this.props.updateSearchParams(params);
-  }
-
-  handleServiceTypeChange = serviceType => {
-    const params = {
-      ...this.props.searchParams,
-      ...this.state.params,
-      text: this.state.text,
-      type: serviceType,
-      priceLevel: undefined,
-      priceStart: undefined,
-      priceEnd: undefined,
-      tags: undefined,
-      sortBy: undefined,
-    };
-    this.props.updateSearchParams(params);
-  };
-
   render() {
-    const { isMobile, toggleSearch, windowLocation } = this.props;
+    const { isMobile } = this.props;
     return (
       <Wrapper isMobile={isMobile} inFocus={this.state.inFocus}>
         <Inner>
@@ -217,29 +69,18 @@ export default class DesktopSearch extends Component {
               <SearchIcon />
             </IconButton>
           </div>
-          <Form onSubmit={this.handleSubmit}>
-            <SemanticLocationControl
-              onChange={this.handleLocationChange}
-              customStyle={suggestionStyle}
-              {...locationProps}
-              defaultAddress={getAddress(this.props.searchParams)}
-              ref={this.inputRef}
-              showServiceTypes
-              handleServiceTypeChange={this.handleServiceTypeChange}
-              serviceType={this.props.searchParams.type}
-              hasSearchedText={Boolean(this.state.text)}
-              isInResultsPage={windowLocation.pathname === '/results'}
-              showGoButton={windowLocation.pathname !== '/results'}
-              updateSearchParams={this.props.updateSearchParams}
-            />
-            {isMobile && (
-              <SubmitButton style={{ color: 'grey' }}>
-                <ArrowWrap onClick={toggleSearch}>
-                  <CrossIcon style={{ color: 'grey' }} />
-                </ArrowWrap>
-              </SubmitButton>
-            )}
-          </Form>
+
+          <Suspense
+            fallback={
+              <input
+                type="text"
+                placeholder="Where would you like to go?"
+                style={{ ...suggestionStyle, border: 'none', 'outline-width': 0 }}
+              />
+            }
+          >
+            <SearchInput {...this.props} />
+          </Suspense>
         </Inner>
       </Wrapper>
     );
